Cut boot JS ~44% by code-splitting heavy desktop shell dependencies
- Lazy-load the weather / lyrics / cover dynamic wallpaper layers so three.js,
  react-player, karaoke visuals, LyricsDisplay (romanization libs), and the
  iPod/Karaoke store stack no longer load at boot on other wallpapers; CSS-only
  fallbacks render while a layer's chunk loads.
- Split the cover-based wallpaper accent into a lazily mounted runner and add a
  tiny now-playing-cover bridge store so the Aqua Glass menubar tone sampler no
  longer statically imports the iPod stores.
- Defer the AirDrop listener to idle (drops the 1800-line finder VFS hook and
  its store graph from the entry chunk).
- Lazy-load DebugLogOverlay behind the Debug Mode flag.
- Make the CJK webfont stylesheets non-render-blocking (print->all swap).
- Type-only imports for @tiptap/core JSONContent and finder SortType.
- Add scripts/trace-import-chain.ts and a boot-import-graph wiring test that
  fails with the offending chain if a heavy module becomes statically
  reachable from src/main.tsx again.

Entry chunk: 1549 kB -> 909 kB (491 kB -> 260 kB gzip); boot no longer
fetches the three (465 kB), media-player, and hangul chunks.
