+++ src/App.tsx
@@ -1,6 +1,6 @@
 import { AppManager } from "./apps/base/AppManager";
 import { appRegistry } from "./config/appRegistry";
-import { useEffect, useMemo, useReducer, useCallback } from "react";
+import { Suspense, lazy, useEffect, useMemo, useReducer, useCallback } from "react";
 import { applyDisplayMode } from "./utils/displayMode";
 import { Toaster } from "./components/ui/sonner";
 import { toast } from "@/hooks/useToast";
@@ -24,12 +24,20 @@ import { ScreenSaverOverlay } from "./components/screensavers/ScreenSaverOverlay
 import { useBackgroundChatNotifications } from "./hooks/useBackgroundChatNotifications";
 import { DesktopErrorBoundary } from "@/components/errors/ErrorBoundaries";
 import { DeferredAutoCloudSync } from "@/hooks/useDeferredAutoCloudSync";
-import { AirDropListener } from "@/components/AirDropListener";
+import { DeferredAirDropListener } from "@/components/DeferredAirDropListener";
 import { WallpaperAccentRunner } from "@/hooks/WallpaperAccentRunner";
 import { DesktopCornerMask } from "@/components/layout/desktop/DesktopCornerMask";
 import { installNativeToastNotifications } from "@/utils/nativeToastNotifications";
-import { DebugLogOverlay } from "@/components/debug/DebugLogOverlay";
 import { createClientLogger } from "@/utils/logger";
+
+// Code-split: the debug overlay (console/network panels + live dashboard) is
+// a large component tree that only renders while Debug Mode is enabled, so
+// normal sessions never download or parse it.
+const DebugLogOverlay = lazy(() =>
+  import("@/components/debug/DebugLogOverlay").then((m) => ({
+    default: m.DebugLogOverlay,
+  }))
+);
 
 // Convert registry to array
 const apps: AnyApp[] = Object.values(appRegistry);
@@ -76,7 +84,10 @@ export function App() {
       setLastSeenDesktopVersion: state.setLastSeenDesktopVersion,
     })
   );
-  const displayMode = useDisplaySettingsStoreShallow((state) => state.displayMode);
+  const { displayMode, debugMode } = useDisplaySettingsStoreShallow((state) => ({
+    displayMode: state.displayMode,
+    debugMode: state.debugMode,
+  }));
   const { isWindowsTheme, isMacOSTheme, isSystem7Theme, isAquaGlass } =
     useThemeFlags();
   const isMobile = useIsMobile();
@@ -273,11 +284,15 @@ export function App() {
         <AppManager apps={apps} />
       </DesktopErrorBoundary>
       <Toaster position={toastConfig.position} offset={toastConfig.offset} />
-      <AirDropListener />
+      <DeferredAirDropListener />
       <DeferredAutoCloudSync />
       <WallpaperAccentRunner />
       <ScreenSaverOverlay />
-      <DebugLogOverlay />
+      {debugMode && (
+        <Suspense fallback={null}>
+          <DebugLogOverlay />
+        </Suspense>
+      )}
     </>
   );
 }
